backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
numpy==2.3.5
Pillow==12.0.0
